+++ droidboy/src/main/java/com/appboy/sample/util/LifecycleUtils.java
@@ -7,6 +7,7 @@ import android.content.Intent;
 
 import com.appboy.sample.activity.DroidBoyActivity;
 import com.appboy.support.AppboyLogger;
+import com.appboy.support.IntentUtils;
 
 @SuppressWarnings("PMD.DoNotCallSystemExit")
 public class LifecycleUtils {
@@ -15,7 +16,8 @@ public class LifecycleUtils {
   public static void restartApp(Context context) {
     Intent startActivity = new Intent(context, DroidBoyActivity.class);
     int pendingIntentId = 109829837;
-    PendingIntent pendingIntent = PendingIntent.getActivity(context, pendingIntentId, startActivity, PendingIntent.FLAG_CANCEL_CURRENT);
+    final int flags = PendingIntent.FLAG_CANCEL_CURRENT | IntentUtils.getDefaultPendingIntentFlags();
+    PendingIntent pendingIntent = PendingIntent.getActivity(context, pendingIntentId, startActivity, flags);
     AlarmManager alarmManager = (AlarmManager) context.getSystemService(Context.ALARM_SERVICE);
     alarmManager.set(AlarmManager.RTC, System.currentTimeMillis() + 100, pendingIntent);
     AppboyLogger.i(TAG, "Restarting application to apply new environment values");