+++ samples/custom-broadcast/src/main/java/com/braze/custombroadcast/CustomBroadcastApplication.java
@@ -1,0 +1,24 @@
+package com.braze.custombroadcast;
+
+import android.app.Application;
+import android.util.Log;
+
+import com.appboy.Appboy;
+import com.appboy.AppboyLifecycleCallbackListener;
+import com.appboy.configuration.AppboyConfig;
+import com.appboy.support.AppboyLogger;
+
+public class CustomBroadcastApplication extends Application {
+  @Override
+  public void onCreate() {
+    super.onCreate();
+    AppboyLogger.setLogLevel(Log.VERBOSE);
+
+    AppboyConfig.Builder appboyConfig = new AppboyConfig.Builder()
+        .setDefaultNotificationChannelName("Appboy Push")
+        .setDefaultNotificationChannelDescription("Appboy related push");
+    Appboy.configure(this, appboyConfig.build());
+
+    registerActivityLifecycleCallbacks(new AppboyLifecycleCallbackListener());
+  }
+}